refactor(structure): move benchmarks, scenarios, adapters, utils to dedicated folders; update imports/project layout
