+++ benchmark_cherrypick/analysis_options.yaml
@@ -15,6 +15,7 @@ include: package:lints/recommended.yaml
 analyzer:
   errors:
     deprecated_member_use: ignore
+    depend_on_referenced_packages: ignore
 
 # Uncomment the following section to specify additional rules.
