+++ benchmark_cherrypick/lib/di_adapters/cherrypick_adapter.dart
@@ -1,12 +1,5 @@
 import 'package:cherrypick/cherrypick.dart';
-
-abstract class DIAdapter {
-  void setupModules(List<Module> modules);
-  T resolve<T>({String? named});
-  Future<T> resolveAsync<T>({String? named});
-  void teardown();
-  DIAdapter openSubScope(String name);
-}
+import 'di_adapter.dart';
 
 class CherrypickDIAdapter implements DIAdapter {
   Scope? _scope;
