+++ benchmark_cherrypick/lib/di_adapters/di_adapter.dart
@@ -1,0 +1,9 @@
+import 'package:cherrypick/cherrypick.dart';
+
+abstract class DIAdapter {
+  void setupModules(List<Module> modules);
+  T resolve<T>({String? named});
+  Future<T> resolveAsync<T>({String? named});
+  void teardown();
+  DIAdapter openSubScope(String name);
+}
