+++ benchmark_cherrypick/lib/scenarios/app_module.dart
@@ -1,0 +1,9 @@
+import 'package:cherrypick/cherrypick.dart';
+import 'foo_service.dart';
+
+class AppModule extends Module {
+  @override
+  void builder(Scope currentScope) {
+    bind<FooService>().toProvide(() => FooService());
+  }
+}
