+++ benchmark_cherrypick/lib/scenarios/child_impl.dart
@@ -1,0 +1,2 @@
+import 'shared.dart';
+class ChildImpl extends Shared {}
