+++ benchmark_cherrypick/lib/scenarios/child_override_module.dart
@@ -1,0 +1,10 @@
+import 'package:cherrypick/cherrypick.dart';
+import 'child_impl.dart';
+import 'shared.dart';
+
+class ChildOverrideModule extends Module {
+  @override
+  void builder(Scope currentScope) {
+    bind<Shared>().toProvide(() => ChildImpl()).singleton();
+  }
+}
