+++ benchmark_cherrypick/lib/scenarios/foo_service.dart
@@ -1,0 +1,1 @@
+class FooService {}
