+++ benchmark_cherrypick/lib/scenarios/parent_impl.dart
@@ -1,0 +1,2 @@
+import 'shared.dart';
+class ParentImpl extends Shared {}
