+++ benchmark_cherrypick/lib/scenarios/parent_module.dart
@@ -1,0 +1,10 @@
+import 'package:cherrypick/cherrypick.dart';
+import 'parent_impl.dart';
+import 'shared.dart';
+
+class ParentModule extends Module {
+  @override
+  void builder(Scope currentScope) {
+    bind<Shared>().toProvide(() => ParentImpl()).singleton();
+  }
+}
