+++ benchmark_cherrypick/lib/scenarios/service.dart
@@ -1,0 +1,5 @@
+abstract class Service {
+  final dynamic value;
+  final Service? dependency;
+  Service({required this.value, this.dependency});
+}
