+++ benchmark_cherrypick/lib/scenarios/service_impl.dart
@@ -1,0 +1,5 @@
+import 'service.dart';
+
+class ServiceImpl extends Service {
+  ServiceImpl({required super.value, super.dependency});
+}
