+++ benchmark_cherrypick/lib/scenarios/shared.dart
@@ -1,0 +1,1 @@
+class Shared {}
